@@ -1,12 +1,11 @@
 module github.com/alexmerren/speedruncom-scraper
 
-go 1.21.0
-
-toolchain go1.21.1
+go 1.23.4
 
 require (
-	github.com/alexmerren/httpcache v0.0.0-20240323204617-c869defe2edb
+	github.com/alexmerren/httpcache v0.0.0-20250118150303-3df901f323d0
+	github.com/alistanis/cartesian v0.0.0-20220409094110-a224e60a7f74
 	github.com/buger/jsonparser v1.1.1
 )
 
-require github.com/mattn/go-sqlite3 v1.14.17 // indirect
+require github.com/mattn/go-sqlite3 v1.14.24 // indirect
